Refactoring the folders and files.
