+++ external/html/index.html
@@ -3,14 +3,14 @@
  <head> 
   <meta name="generator" content="HTML Tidy for Linux (vers 25 March 2009), see www.w3.org" /> 
   <script src="http://code.jquery.com/jquery-2.1.4.min.js" type="text/javascript"></script>
-  <script src="http://192.168.199.10:8080/js/base64.js" type="text/javascript"></script> 
+  <script src="https://containerops.me/js/base64.js" type="text/javascript"></script> 
   <title>websokcet test</title> 
  </head> 
  <body> 
 <script type = "text/javascript">
     var time = 1000;
     var sock;
-    var wsuri = "ws://192.168.19.111:8080/ws";
+    var wsuri = "ws://containerops.me/ws";
 $(function() {
 
     try {
