+++ handler/build.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 
 	"github.com/Unknwon/macaron"
+
 	"github.com/containerops/generator/models"
 	. "github.com/containerops/generator/modules"
 	"github.com/containerops/wrench/utils"
